% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{convert_currency}
\alias{convert_currency}
\title{Get exchange rates between two currencies in a given time frame}
\usage{
convert_currency(to, from, n_days, retried = 0)
}
\arguments{
\item{to}{Currency to convert TO}

\item{from}{Currency to convert FROM}

\item{n_days}{Number of days from current date}

\item{retried}{Number of retries in the previous step}
}
\description{
Hitting the exchangerates.io API to get historical exchange rate from today n days backwards
}
