% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{euro}
\alias{euro}
\title{Format a number with the EUR symbol}
\usage{
euro(x)
}
\arguments{
\item{x}{number}
}
\value{
string with the EUR symbol as prefix
}
\description{
Format a number with the EUR symbol
}
\examples{
euro(42)
euro(4.2*1e7)
}
