% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{forint}
\alias{forint}
\title{Function to format currency to HUF}
\usage{
forint(amt)
}
\arguments{
\item{amt}{Numeric currency amount}
}
\value{
Formatted currency.
}
\description{
Function to format currency to HUF
}
