% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Binance API to get most recent USD price of BTC}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{the number of retries previously done}
}
\description{
Binance API to get most recent USD price of BTC
}
