% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btc_value.R
\name{get_btcusdt}
\alias{get_btcusdt}
\title{Retrieve most recent BTC value in USD}
\usage{
get_btcusdt(attempt = 1)
}
\arguments{
\item{attempt}{1}
}
\value{
num value of 1 BTC in USD
}
\description{
Retrieve most recent BTC value in USD
}
\examples{
get_btcusdt()
}
