% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rates.R
\name{get_exchange_rate}
\alias{get_exchange_rate}
\title{Getting exchange rate between USD and HUF for 30 days}
\usage{
get_exchange_rate(
  base_currency,
  end_currency,
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{base_currency}{character}

\item{end_currency}{character}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function has failed}
}
\value{
\code{data.table} object
}
\description{
Getting exchange rate between USD and HUF for 30 days
}
\examples{
get_exchange_rate("USD", "HUF")
}
