% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rates.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{This function returns exchange rates from any two currencies at the given period}
\usage{
get_exchange_rates(
  currency1 = "USD",
  currency2 = "EUR",
  initial_date = Sys.Date() - 30,
  final_date = Sys.Date()
)
}
\arguments{
\item{currency1}{first currency, default USD}

\item{currency2}{2nd currency, default EUR}

\item{initial_date}{initial date, default 30 days ago}

\item{final_date}{final date, default EUR}
}
\description{
This function returns exchange rates from any two currencies at the given period
}
