% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{get_usdeur}
\alias{get_usdeur}
\title{Retrieve USD/EUR exchange rate using API}
\usage{
get_usdeur()
}
\value{
num value of 1 USD in EUR
}
\description{
Retrieve USD/EUR exchange rate using API
}
\examples{
get_usdeur()
}
