% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhuf.R
\name{get_usdhuf}
\alias{get_usdhuf}
\title{Getting exchange rate between USD and HUF}
\usage{
get_usdhuf(retried = 0)
}
\arguments{
\item{retried}{number of times the function has failed}
}
\value{
number
}
\description{
Getting exchange rate between USD and HUF
}
\examples{
get_usdhuf()
}
