% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usdhuf.R
\name{get_usdhuf_2}
\alias{get_usdhuf_2}
\title{Getting exchange rate between USD and HUF for 30 days}
\usage{
get_usdhuf_2(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}

\item{retried}{number of times the function has failed}
}
\value{
\code{data.table} object
}
\description{
Getting exchange rate between USD and HUF for 30 days
}
\examples{
get_usdhuf_2()
}
