\name{mr-package}
\alias{mr-package}
\docType{package}
\title{
Marginal Regresson Models for Dependent Data
}
\description{
 Fits marginal regression models described by Masarotto and Varin (2010). 
}
\details{
This package contains \R functions, C code and data sets related to the paper Masarotto and Varin (2010). The main function is \code{\link{mr}}. This allows for likelihood inference in a class of marginal regression models for non-normal responses with a flexible specification of the dependence structure.   
}
\author{
Guido Masarotto and Cristiano Varin.
}
\references{
Masarotto, G. and Varin, C. (2010). Gaussian dependence models for non-Gaussian marginal regression. Preprint. 
}
\keyword{package}
