\name{mrind}
\alias{mrind}
\title{
Working independence errors in Marginal Regression Models
}
\description{
Sets working independence errors in marginal regression models.
}
\usage{
mrind()
}
\value{
An object of class \code{\link{cor.mr}} representing an identity correlation matrix.
}
\references{
Masarotto, G. and Varin, C. (2010). Gaussian dependence models for non-Gaussian marginal regression. Preprint.
}
\author{
Guido Masarotto and Cristiano Varin.
}
\seealso{
\code{\link{mr}}
}
\keyword{regression}
\keyword{nonlinear}