\name{polio}
\alias{polio}
\alias{y.polio}
\alias{x.polio}
\docType{data}
\title{
Polio Time Series 
}
\description{
Time series of Polio incidences in U.S.A. from 1970 to 1983.
}
\usage{data(polio)}
\format{
\tabular{ll}{
\code{y.polio} \tab time series of polio incidences. \cr
\code{x.polio} \tab model matrix with intercept, linear trend and seasonal components. \cr}
}
\source{
Zeger, S.L. (1988). A regression models for time series of counts. \emph{Biometrika} \bold{75}, 822--835.
}
\examples{
data(polio)
}
\keyword{datasets}
