% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{query_rates}
\alias{query_rates}
\title{Hitting the exchangerates.io API to get historical exchange rates between two currencies
in a given time period}
\usage{
query_rates(to, from, n_days, retried = 0)
}
\arguments{
\item{to}{Currency to convert TO}

\item{from}{Currency to convert FROM}

\item{n_days}{Number of days from current date}

\item{retried}{Number of retries in the previous step}
}
\description{
Hitting the exchangerates.io API to get historical exchange rates between two currencies
in a given time period
}
