\name{salamanders}
\alias{salamanders}
\docType{data}
\title{
Salamander Mating Data  
}
\description{
Data from a study on barriers to interbreeding in two geographically isolated species of salamanders called Whiteside and Rough-Butt. 
}
\usage{data(salamanders)}
\format{
\tabular{ll}{
\code{y1} \tab binary responses from the first experiment in Summer 1986. \cr
\code{y2} \tab binary responses from the second experiment in Fall 1986. \cr
\code{y3} \tab binary responses from the third experiment in Fall 1986. \cr
\code{x} \tab model matrix. \cr
\code{zf} \tab matrix with indicators of which female of salamander is involved in each mating experiment. \cr
\code{zm} \tab matrix with indicators of which male of salamander is involved in each mating experiment.
}
}
\source{
McCullagh, P. and Nelder, J.A. (1989). \emph{Generalized Linear Models}. Second edition. Chapman and Hall.
}
\examples{
data(salamanders)
}
\keyword{datasets}
