% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{convert_currency}
\alias{convert_currency}
\title{Get exchange rates for a pair of currencies for a certain number of days}
\usage{
convert_currency(base, quote, days)
}
\arguments{
\item{base:}{base currency}

\item{quote:}{quote currency}

\item{days:}{number of days to return}
}
\value{
Table that indicates how much of the \code{quote} currency is needed to purchase one unit of
the \code{base} currency for a given number of \code{days}
}
\description{
Get exchange rates for a pair of currencies for a certain number of days
}
\examples{
convert_currency("USD", "JPY", 10)
convert_currency("EUR", "HUF", 15)
}
