% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{convert_currency}
\alias{convert_currency}
\title{Get exchange rates for a pair of currencies for a certain number of days}
\usage{
convert_currency(base, symbol, days)
}
\arguments{
\item{base:}{base currency, symbol: equivalent in a given currency, days: number of days}
}
\description{
Get exchange rates for a pair of currencies for a certain number of days
}
