% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{convert_currency}
\alias{convert_currency}
\title{Queries historical exchange rates for any currency (so configurable symbol and base currency) for the past number of days}
\usage{
convert_currency(curr_from, curr_to, days)
}
\arguments{
\item{curr_from}{string}

\item{curr_to}{string}

\item{days}{number}
}
\description{
Queries historical exchange rates for any currency (so configurable symbol and base currency) for the past number of days
}
