% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{convert_currency}
\alias{convert_currency}
\title{Query historical exchange rates}
\usage{
convert_currency(base, symbol, days)
}
\arguments{
\item{base}{base currency}

\item{symbol}{currency to convert to}

\item{days}{number of days to go back from current date}
}
\description{
Query historical exchange rates for any currency for the past number of days from https://exchangeratesapi.io/.
}
\examples{
convert_currency('EUR', 'HUF', 10)
}
