% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_currency_converter.R
\name{convert_currency}
\alias{convert_currency}
\title{Function to query historical exchange rates for any currency for the given number of days in the past}
\usage{
convert_currency(base, symbol, days)
}
\arguments{
\item{base}{is the currency we want to convert}

\item{symbol}{is the currency to which we want to convert the base currency}

\item{days}{is the number of days to retroactively query the exchange rates}
}
\description{
Function to query historical exchange rates for any currency for the given number of days in the past
}
