% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{convert_currency}
\alias{convert_currency}
\title{Function to retrieve historical exchange rates of a base to a symbol currency for the
past x amount of days}
\usage{
convert_currency(from, to, days)
}
\arguments{
\item{from}{string - base currency}

\item{to}{string - target currency}

\item{days}{int - T-number of days we're querying}
}
\value{
data.table
}
\description{
Function to retrieve historical exchange rates of a base to a symbol currency for the
past x amount of days
}
