% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{euro}
\alias{euro}
\title{Format a number with the Euro symbol}
\usage{
euro(x)
}
\arguments{
\item{x}{number}
}
\value{
string with the Euro symbol prefixed
}
\description{
Format a number with the Euro symbol
}
\examples{
euro(42)
euro(420000000)
}
