% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{euro}
\alias{euro}
\title{format a number with the euro symbol}
\usage{
euro(x)
}
\arguments{
\item{x}{number}
}
\value{
string with the euro symbol prefixed
}
\description{
format a number with the euro symbol
}
\examples{
euro(42)
euro(42000000)
}
