% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Formatter function for Hungarian Forint}
\usage{
forint(number)
}
\arguments{
\item{number:}{converts a given number to the forint format}
}
\description{
Formatter function for Hungarian Forint
}
\examples{
forint(42)
}
