% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Convert USD to HUF}
\usage{
forint(x)
}
\arguments{
\item{x}{amount in USD}
}
\value{
string
}
\description{
Convert USD to HUF
}
\examples{
forint(100000)
forint(10.3241245125125)
forint(42)
}
