% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{This function adds HUF at the end of the value passed as argument}
\usage{
forint(x)
}
\arguments{
\item{numeric}{value to be formatted}
}
\description{
This function adds HUF at the end of the value passed as argument
}
