% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Function to add the appropriate suffix to a value in HUF}
\usage{
forint(x)
}
\arguments{
\item{x}{is the value in HUF we would like to add the Ft suffix to}
}
\description{
Function to add the appropriate suffix to a value in HUF
}
