% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Access Binance API for BTC price in USD}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{# the number of retries before sleep}
}
\value{
Return the latest price of 1 BTC in USD
}
\description{
Access Binance API for BTC price in USD
}
