% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{get the most recent BTC price from Binance API. If error, run again}
\usage{
get_bitcoin_price()
}
\arguments{
\item{retried}{the number if retries previously done before the exponential backoff sleep}
}
\description{
This is a wrapper around the \code{binancer} package ...
}
