% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Function to get current bitcoin price from binancer}
\usage{
get_bitcoin_price(retried = 0)
}
\description{
Function to get current bitcoin price from binancer
}
