% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bitcoin_price.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{This returns the current price of bitcoins in dollars}
\usage{
get_bitcoin_price()
}
\description{
This returns the current price of bitcoins in dollars
}
\examples{
get_bitcoin_price()
}
