% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Function to get the price of one BTC in USD by hitting the Binance API'}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{the number of retries previously done before the exponential backoff sleep}
}
\description{
This function is using the \code{binance_coins_prices} function from the \code{binancer} package.
This is a wrapper around the \code{binancer} package.
}
