% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Hitting the Binance API to get the most recent price of a Bitcoin in USD}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{the number of retires previously done before the exponential backoff sleep}
}
\description{
This is a wrapper around the \code{binancer} package ...
}
